requirements.txt
anthropic==0.74.1
python-dotenv==1.2.1
